move creating action fingerprints to domain
